pyproject.toml
[tool.poetry]
name = "rldoom"
version = "0.1.0"
description = "Python package "
license = "MIT"
authors = ["Yanick Medina <yanickjmedina@gmail.com>"]
maintainers = ["Yanick Medina <yanickjmedina@gmail.com>"]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{ include = "rldoom" }]

[tool.poetry.dependencies]
python = "^3.10"
vizdoom = "^1.2.4"
opencv-python = "^4.11.0.86"
gymnasium = "^1.1.1"
matplotlib = ">=3"
pandas = "^2.2.3"
torch = "^2.6.0"
pygame = "^2.6.1"
torchvision = "^0.21.0"
torchtyping = "^0.1.5"
tqdm = "^4.67.1"
typer = "^0.15.2"
PyYAML = "^6.0.2"
pydantic = "^2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"